// src/api/generated/model/rankingUser.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * 주머니(Jumoney) API 문서
 * 주머니의 백엔드 API 명세서입니다.
 * OpenAPI spec version: 1.0.0
 */
import type { TopHolding } from './topHolding';

export interface RankingUser {
  rank?: number;
  userId?: number;
  masterId?: number;
  nickname?: string;
  totalAsset?: number;
  totalProfitRate?: number;
  representativeStocks?: TopHolding[];
}